fix: Fix custom pre-commit hook
This command was being run by pre-commit, as, for example, `pre-commit run nbqa --all-files` or `--files ...`

Given the configuration in .pre-commit-config.yaml, this runs: `nbqa manage FILE ...`

nbqa creates a subprocess, replacing the original files with temporary files. (It then rewrites stdout and stderr to
replace the temporary filenames with the original filenames, making this behavior hard to observe.)

The temporary files are transformations of the original files, in which magic cells are removed. This makes sense for
other linters, that operate on Python code.

Since we want to operate on SQL magics, nbqa is an inappropriate entry point for a pre-commit hook.